This PR is one in a series of PRs created due to benchmarking and improving the `LOAD CSV` clause, which was tested and documented [here](https://github.com/memgraph/memgraph/pull/753).

Here I am implementing `ReferenceBasedEvaluator` as it was shown that copying of array of values slows down performance when there are many properties in a `row` created when reading a CSV file.

[master < Task] PR
- [x] Check, and update documentation if necessary
- [ ] Provide the full content or a guide for the final git message

**Release note**: Implemented `ReferenceBasedEvaluator` as it was shown that copying of an array of values slows down performance when there are many properties in a `row` created when reading a CSV file

**Benchmarking**
- [x] Import 1 000 000 lines, from each line create 2 nodes and 1 new edge. Each line will have X properties, of 10 string length. Each node has X/2 features, where X is even.


| Number of properties per line | Import time (OLD) [seconds] | Import time (NEW) [seconds] | Throughput (OLD) [lines/s] | Throughput (NEW) [lines/s] |
| -- | -- | -- |  -- | -- |
|2 | 3.6s | 3.429 s|  |  |
|10 | 9.6s | 7.962 s|   |  |
|20 | 22s | 14.718 s|  |   |