@vpavicic 2023-02-17
OK but what can i actually do with this? how did this help with anything?

Probably a completely wrong sentence but we need something along the lines of this:

The ReferenceBasedEvaluator helps improve the importing speed of CSV files with more than 10 properties in each row.

The ReferneceBasedEvaulator improves the import of a CSV file ... HOW? Why? :