@antoniofilipovic 2023-02-17
> @antoniofilipovic OK but what can i actually do with this? how did this help with anything?
> 
> Probably a completely wrong sentence but we need something along the lines of this:
> 
> The ReferenceBasedEvaluator helps improve the importing speed of CSV files with more than 10 properties in each row.
> 
> The ReferneceBasedEvaulator improves the import of a CSV file ... HOW? Why? :

I am sorry if this didn't help you, thought it made sense :smile: 
I will describe to you what was slow and suggest one more description for the changelog, maybe it will help more. Firstly, when using Memgraph, by doing profiling I found that performance degrades if users are referencing arrays with a lot of properties, for example, let's say you have an array of the following properties on every node:
`[1,2, ..., 500]`.
And assume you are doing some operations and getting values from the array, in the following way:
`node.big_array[<index>]`. It was shown on profiling that this was slow. 

New idea: By implementing ReferenceBasedEvaluator, users can expect improved performance when accessing values from large arrays or maps with numerous properties.