@vpavicic 2023-02-20
That one is much better!

(Remember that I, just like our user, have very little knowledge about the engine, the under the hood stuff, the deep source of issues and how it's being fixed... so we have 2 lvls of user information -> 1. What is now better/changed?