@antoniofilipovic 2023-02-20
@vpavicic they don't need to know about ReferenceBasedEvaluator :) so we can just say:
`Improved performance when accessing values from large arrays or maps with numerous properties`